+++ Examples/CQRS/Examples.Cqrs/Commands/SimplePostCommandDecorator.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using APIBlox.AspNetCore.Contracts;
 using APIBlox.AspNetCore.Types;
+using APIBlox.NetCore.Contracts;
 using Examples.Resources;
 
 namespace Examples.Commands