+++ Examples/CQRS/Examples.Cqrs/Queries/RequiresInputQueryQueryHandler.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using APIBlox.AspNetCore.Contracts;
 using APIBlox.NetCore.Attributes;
+using APIBlox.NetCore.Contracts;
 
 namespace Examples.Queries
 {